Don't try to reconcile CRs in all namespaces
Before this change, we'd try to reconcile a CR in any namespace, and
then fail if it wasn't in a namespace that we expect. Failure wasn't
handled very well, which led to a persistent error loop in the
relevant controller. This changes that so that the controller doesn't
even bother calling the Reconcile method if the CR isn't in one of the
namespaces that we expect for that kind.

Since running the operator locally with `operator-sdk up local`
doesn't seem to correctly react to CRs in namespaces other than one
for rthe current context, this change also adds that to the list of
`APP_NAMESPACES` for the `make code/run` target.